+++ test/facets/L2/PerpetualMint/behaviors/_selectERC1155Owner.t.sol
@@ -15,7 +15,7 @@ contract PerpetualMint_selectERC1155Owner is
     L2ForkTest
 {
     /// @dev value of roll which will lead to depositor one being selected
-    uint128 internal constant depositorOneSelectValue = uint64(900);
+    uint256 internal constant depositorOneSelectValue = uint256(900);
 
     function setUp() public override {
         super.setUp();
@@ -24,14 +24,14 @@ contract PerpetualMint_selectERC1155Owner is
     }
 
     /// @dev tests selecting an ERC1155 owner after an ERC1155 asset has been won
-    function testFuzz_selectERC1155Owner(uint64 randomOwnerValue) public view {
+    function testFuzz_selectERC1155Owner(uint256 randomOwnerValue) public view {
         uint256 selectedTokenId = PARALLEL_ALPHA_TOKEN_ID_ONE;
 
         // make sure random value is within 0 - tokenRisk range
-        uint64 normalizedValue = randomOwnerValue %
+        uint256 normalizedValue = randomOwnerValue %
             _tokenRisk(address(perpetualMint), PARALLEL_ALPHA, selectedTokenId);
 
-        uint64 depositorOneRisk = _depositorTokenRisk(
+        uint256 depositorOneRisk = _depositorTokenRisk(
             address(perpetualMint),
             depositorOne,
             PARALLEL_ALPHA,
